requirements.txt
chromadb>=0.6.3
docx2txt>=0.8
gradio>=4.44.0,<5.0.0
huggingface_hub<1.0
langchain-core>=0.3.43
langchain-text-splitters>=0.3.6
openai>=1.68.2
pypdf>=5.3.0
python-dotenv>=1.0.1
